NODE-1088: Clean up
