+++ integration-testing/casperlabs_local_net/casperlabs_accounts.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import base64
 
 from casperlabs_local_net.common import resources_path
-from casperlabs_client import read_pem_key
+from casperlabs_client.crypto import read_pem_key
 
 
 def is_valid_account(account_id: Union[int, str]) -> bool:
