+++ integration-testing/client/CasperLabsClient/casperlabs_client/crypto.py
@@ -6,6 +6,19 @@ from cryptography.hazmat.primitives import hashes
 from cryptography.x509.oid import NameOID
 from cryptography.hazmat.primitives import serialization
 from Crypto.Hash import keccak
+from pyblake2 import blake2b
+import ed25519
+from . import consensus_pb2 as consensus
+import base64
+
+base64_b64decode = base64.b64decode
+
+
+def read_pem_key(file_name: str):
+    with open(file_name) as f:
+        s = [l for l in f.readlines() if l and not l.startswith("-----")][0].strip()
+        r = base64_b64decode(s)
+        return len(r) % 32 == 0 and r[:32] or r[-32:]
 
 
 def generate_key_pair():
@@ -60,3 +73,20 @@ def generate_certificates(private_key, public_key):
         encryption_algorithm=serialization.NoEncryption(),
     )
     return cert_pem, key_pem
+
+
+def blake2b_hash(data: bytes) -> bytes:
+    h = blake2b(digest_size=32)
+    h.update(data)
+    return h.digest()
+
+
+def signature(private_key, data: bytes):
+    return private_key and consensus.Signature(
+        sig_algorithm="ed25519",
+        sig=ed25519.SigningKey(read_pem_key(private_key)).sign(data),
+    )
+
+
+def private_to_public_key(private_key) -> bytes:
+    return ed25519.SigningKey(read_pem_key(private_key)).get_verifying_key().to_bytes()
